MNT: Factor out individual CCSDS packet handling/iterating from the parsing
Create a separate CCSDS packet generator that is dedicated to handling the input
data and yielding individual packets. The parsing functions then utilize
this generator to handle the actual data parsing.
